Lesson2/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
